requirements.txt
Flask==3.1.0
Flask-PyMongo==3.0.1
pymongo==4.11.1
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
email-validator==2.1.0.post1
PyJWT==2.8.0
bcrypt==4.1.2
Werkzeug>=3.0.0
gunicorn==21.2.0
